+++ lib/twenv/commands/write_tweet.rb
@@ -1,6 +1,6 @@
 class TWEnv::WriteTweet < TWEnv::Command
   match 'write-tweet'
-  description "Write a tweet"
+  description "Write a tweet on behalf of `client.user`"
   group 'twenv'
   command_options storage: true
   banner <<-BANNER